Addition to description of five-guess algorithm
The description of the five-guess algorithm does not currently state that in the case of multiple codes producing the same guaranteed number of eliminations, one still in the set S should be chosen. This is a requirement of the algorithm as stated by Knuth in his paper (http://www.cs.uni.edu/~wallingf/teaching/cs3530/resources/knuth-mastermind.pdf) - 'If this minimum can be achieved by a "valid" pattern, a valid one should be used'.

Implementing the algorithm as currently stated (not necessarily choosing from codes left in S when there are multiple options) leads to an *almost* working solution which passes the unit tests in most cases, however fails to guess within 5 moves for a small number of codes ("eaea", "eaca", "eaaa" are examples of these).